fix(api,ui): remove hacky personal teams construct (#953)
Teams were kind of tacked on and anything without a team id, aka null, was considered "personal" and only accessible to the user.

This PR changes it so that users get assigned a team we call personal and then we can have everything foreign key to a team.

+++ backend/recipeyak/api/team_detail_view.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from django.db import transaction
 from django.db.models import QuerySet
 from django.shortcuts import get_object_or_404
 from rest_framework import status
@@ -16,9 +17,7 @@ from recipeyak.models.user import User
 
 
 def get_teams(user: User) -> QuerySet[Team]:
-    return Team.objects.filter(membership__user_id=user.id) | Team.objects.filter(
-        is_public=True
-    )
+    return Team.objects.filter(membership__user_id=user.id)
 
 
 def is_team_admin(team: Team, user: User) -> bool:
@@ -27,7 +26,7 @@ def is_team_admin(team: Team, user: User) -> bool:
 
 @api_view(["DELETE", "PATCH", "GET"])
 @permission_classes([IsAuthenticated])
-def team_detail_view(request: AuthedRequest, team_pk: str) -> Response:
+def team_detail_view(request: AuthedRequest, team_pk: int) -> Response:
     if request.method == "PATCH":
         team = get_object_or_404(get_teams(request.user), pk=team_pk)
         if not is_team_admin(team, request.user):
@@ -44,10 +43,15 @@ def team_detail_view(request: AuthedRequest, team_pk: str) -> Response:
         serializer = TeamSerializer(team)
         return Response(serializer.data)
     elif request.method == "DELETE":
-        team = get_object_or_404(get_teams(request.user), pk=team_pk)
-        if not is_team_admin(team, request.user):
-            return Response(status=403)
-        team.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        with transaction.atomic():
+            team = get_object_or_404(get_teams(request.user), pk=team_pk)
+            if (
+                not is_team_admin(team, request.user)
+                # don't allow deleting last team
+                or get_teams(request.user).count() <= 1
+            ):
+                return Response(status=403)
+            team.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
     else:
         raise MethodNotAllowed(request.method or "")
